rename the package name and delete comments
