+++ pkg/scaling/scaledjob/metrics_test.go
@@ -1,4 +1,4 @@
-package utils
+package scaledjob
 
 import (
 	"testing"
